+++ src/WebApiBolierplate/Core.Constants/JWT.cs
@@ -1,9 +1,0 @@
-﻿namespace Core.Constants
-{
-    public static class JWT
-    {
-        public const string Key = "This is a sample key";
-        public const string Issuer = "self";
-        public const string Audience = "https://www.example.com";
-    }
-}