+++ src/WebApiBolierplate/Core.Constants/Roles.cs
@@ -1,0 +1,8 @@
+﻿namespace Core.Constants
+{
+    public static class Roles
+    {
+        public const string Admin = "admin";
+        public const string User = "user";
+    }
+}