+++ src/WebApiBolierplate/Model/LoginDTO.cs
@@ -1,0 +1,13 @@
+﻿namespace Model
+{
+    public class LoginDTO
+    {
+        public LoginDTO()
+        {
+
+        }
+
+        public string UserName { get; set; }
+        public string Password { get; set; }
+    }
+}