+++ src/WebApiBolierplate/Model/User.cs
@@ -1,0 +1,16 @@
+﻿namespace Model
+{
+    public class User
+    {
+        public User()
+        {
+
+        }
+
+        public string Email { get; set; }
+        public string Name { get; set; }
+        public string CompanyId { get; set; }
+        public string Roles { get; set; }
+        public string Id { get; set; }
+    }
+}