+++ CHANGELOG.md
@@ -4,13 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.2] - 2019-01-10
+### Added
+- Completed JWT token generation.
+- Created business library.
+- Added postman collection file.
+
+### Changed
+- Removed Tuple declarations.
+- Constants are now added in statup.cs.
+- Other minor code optimizations.
 
 ## [0.0.1] - 2019-01-02
 ### Added
 - Added changelog.
 - Added core library.
-- Added API Helper functions
+- Added API Helper functions.
 
 ### Changed
-- Format of model validation errors
+- Format of model validation errors.